Rename directory "ifs" to "interfaces"
Fixes devlights/try-golang#276
