--- examples/basic/ifs/doc.go
@@ -1,4 +1,0 @@
-/*
-Package ifs -- Go言語のインターフェースについてのサンプルが配置されているパッケージです。
-*/
-package ifs
