+++ examples/basic/interfaces/command/command.go
