+++ examples/basic/interfaces/doc.go
@@ -1,0 +1,4 @@
+/*
+Package interfaces -- Go言語のインターフェースについてのサンプルが配置されているパッケージです。
+*/
+package interfaces
