+++ examples/basic/interfaces/interface_basic.go
@@ -1,4 +1,4 @@
-package ifs
+package interfaces
 
 import "fmt"
