+++ examples/basic/interfaces/interface_ducktyping.go
@@ -1,4 +1,4 @@
-package ifs
+package interfaces
 
 import "fmt"
